refactor: 나의 문제집 페이지 관련 type이 문제풀이 페이지 types에 들어있어서 분리시켜줌
