+++ src/features/solve/types/question.ts
@@ -22,13 +22,3 @@ export interface QuestionSet {
   updateTime: string; // ISO 8601 문자열 (예: 2025-09-16T23:41:32.129409)
 }
 
-export interface MyQuestionSetsResponse {
-  questionSetId: number;
-  title: string;
-  sourceIds: number[];
-  sourceNames: string[];
-  questionCount: number;
-  difficultyType: DifficultyType;
-  questionType: QuestionType;
-  createdAt: string;
-}
